requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.9
python-telegram-bot==20.7
python-dotenv==1.0.0
pillow==10.1.0
redis==5.0.1
celery==5.3.4
bcrypt==4.1.2
requests==2.31.0
gunicorn==21.2.0
stripe==7.8.0
